write-behind/pubsub-sql/functions/requirements.txt
google-cloud-pubsub==2.13.9
psycopg2-binary==2.9.5
